requirements.txt
Flask
Flask-Session
Flask-Cors
python-dotenv
spotipy
numpy
aiohttp
requests
